refactor(random-mix): G.88.1 — extract helpers + AUDIOBOOK_GENRES + filter logic
